cleanup

+++ a-sysmon-core/src/main/java/com/ajjpj/asysmon/data/AGlobalDataPoint.java
@@ -7,10 +7,10 @@ public class AGlobalDataPoint {
     private static double[] pow = new double[] {1.0, 10.0, 100.0, 1000.0, 10*1000.0, 100*1000.0, 1000*1000.0, 10*1000*1000.0, 100*1000*1000.0, 1000*1000*1000.0};
 
     private final String name;
-    private final int value;
+    private final long value;
     private final int numFracDigits;
 
-    public AGlobalDataPoint(String name, int value, int numFracDigits) {
+    public AGlobalDataPoint(String name, long value, int numFracDigits) {
         this.name = name;
         this.value = value;
         this.numFracDigits = numFracDigits;
